* iOS_concurrent:
  Remove unused code lines
  Clean up debugging text output
  Fix bugs of 'wandering' frame-rate i.e. actual frame-rate is different from the fixed constant
  Fix bugs with frame-block numbering
  Update code-signature
  Refactor codes for auto-stop
  Fixed fingerLabel does not show up
  Implemented autostop for finger and face
  Factor code for finger auto-start into auto_start.mm
  Complete auto-start for finger
  refactor code
  fix some bugs
  Implemented autostart for finger
  Correct spelling errors in the variable names
  Reorder documentation

Conflicts:
	MisfitHeartRate/Pulsar.xcodeproj/project.xcworkspace/xcuserdata/misfit.xcuserdatad/UserInterfaceState.xcuserstate

+++ MisfitHeartRate/MisfitHeartRate/algorithms/config.h
@@ -39,7 +39,7 @@ namespace MHR {
     extern double _eulerian_chromaMagnifier; // Standard: 1
     
     // Native params of the algorithm
-    extern int _frameRate;
+    extern double _frameRate;
     extern int _number_of_channels;
     extern int _Gpyr_filter_length;
     extern int _startFrame;
